requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
folium>=0.14.0
geopandas>=0.13.0
streamlit>=1.25.0
google-api-python-client>=2.88.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.10.0
scikit-learn>=1.3.0 